requirements.txt
numpy
torch
torchvision
Pillow
scikit-learn
opencv-python
nltk